dynamic-web-server/stream/requirements.txt
pandas
tornado
websockets
websocket-client
perspective-python==3.4.3
# ^3.5.x is broken